requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt==3.2.2
passlib==1.7.4
python-jose
python-multipart
